+++ src/main/java/org/ivan/artshow/module/comment/pojo/Comment.java
@@ -23,10 +23,10 @@ public class Comment {
     private Long commentId;
 
     @Column(name = "post_id")
-    private Integer postId;
+    private Long postId;
 
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "content")
     private String content;
@@ -45,19 +45,19 @@ public class Comment {
         this.commentId = commentId;
     }
 
-    public Integer getPostId() {
+    public Long getPostId() {
         return postId;
     }
 
-    public void setPostId(Integer postId) {
+    public void setPostId(Long postId) {
         this.postId = postId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
