+++ src/main/java/org/ivan/artshow/module/post/pojo/Post.java
@@ -23,10 +23,10 @@ public class Post {
     private Long postId;
 
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "subcategory_id")
-    private Integer subcategoryId;
+    private Long subcategoryId;
 
     @Column(name = "title")
     private String title;
@@ -51,19 +51,19 @@ public class Post {
         this.postId = postId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getSubcategoryId() {
+    public Long getSubcategoryId() {
         return subcategoryId;
     }
 
-    public void setSubcategoryId(Integer subcategoryId) {
+    public void setSubcategoryId(Long subcategoryId) {
         this.subcategoryId = subcategoryId;
     }
